삭제 시 권한 확인 추가

+++ demo/src/main/java/com/example/demo/board/controller/BoardController.java
@@ -58,7 +58,10 @@ public class BoardController {
 	}
 	
 	@DeleteMapping("/post/{id}")
-	public String delete(@PathVariable("id") Long id, Model model) {
+	public String delete(@PathVariable("id") Long id, Model model, HttpSession session) {
+		if(!session.getAttribute("userId").equals(boardService.get(id).getUserId())) {
+			return "redirect:/denied";
+		}
 		boardService.delete(id);
 		return "redirect:/oauth/board";
 	}
